completed aircraft and airport service implementations

+++ src/main/java/com/aerotravel/flightticketbooking/services/AirportService.java
@@ -11,5 +11,5 @@ public interface AirportService {
     public abstract List<Airport> getAllAirports();
     public abstract Airport getAirportById(Integer airportId);
     public abstract Airport saveAirport(Airport airport);
-    public abstract void deleteAirport(Long airpportId);
+    public abstract void deleteAirport(Integer airpportId);
 }
